tests: Sharper tests. TODO: fix debouncer.
The issue with debouncer - it uses a setTimeout() for queueing microtasks, so they don't block the main thread. THe problem here is, we have zero control or visibility into when they're actually executed AND we can't return values from setTimeouts(), so we can never 100% know when the execution inside debouncer() completed.